implementing 'actAndOpen' mechanism.
The idea would be to allow glamour to encapsulate the action of opening a new glamour browser (the use case i have in mind for now is to open a GLMWizard). 
The model store those sepecific actions (like actions , selectionActions and now openBrowserActions and selectionOpenBrowserActions) , then each renderer now how to deal with them.
What I did for now is to provide a similar interface than for other actions:
aBrowser actAndOpen: [ ... ] entitled: '..'.
The requirement is that the block return a browser.

For now selectionActAndOpen: works fine